+++ services/web/server/src/simcore_service_webserver/security/_authz_policy.py
@@ -13,7 +13,7 @@ from ..db.plugin import get_database_engine
 from ._authz_access_model import OptionalContext, RoleBasedAccessModel, check_access
 from ._authz_db import AuthInfoDict, get_active_user_or_none
 from ._constants import MSG_AUTH_NOT_AVAILABLE
-from ._identity import IdentityStr
+from ._identity_api import IdentityStr
 
 _logger = logging.getLogger(__name__)
 
@@ -80,7 +80,7 @@ class AuthorizationPolicy(AbstractAuthorizationPolicy):
         """Implements Interface: Determines whether an identified user has permission
 
         :param identity: session identified corresponds to the user's email as defined in login.handlers_registration
-        :param permission: name of the operation that user wants to execute OR a tuple as (operator.and_|operator.or_, name1, name2, ...)
+        :param permission: name of the operation that user wants to execute
         :param context: context of the operation, defaults to None
         :return: True if user has permission to execute this operation within the given context
         """
