+++ services/web/server/src/simcore_service_webserver/security/_identity_policy.py
@@ -1,0 +1,5 @@
+from aiohttp_security.session_identity import SessionIdentityPolicy
+
+assert SessionIdentityPolicy  # nosec
+
+__all__: tuple[str, ...] = ("SessionIdentityPolicy",)
