+++ services/web/server/src/simcore_service_webserver/security/plugin.py
@@ -10,13 +10,13 @@ import logging
 
 import aiohttp_security
 from aiohttp import web
-from aiohttp_security.session_identity import SessionIdentityPolicy
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
 from ..session.plugin import setup_session
-from ._authz import AuthorizationPolicy
 from ._authz_access_model import RoleBasedAccessModel
 from ._authz_access_roles import ROLES_PERMISSIONS
+from ._authz_policy import AuthorizationPolicy
+from ._identity_policy import SessionIdentityPolicy
 
 _logger = logging.getLogger(__name__)
